backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
clip @ git+https://github.com/openai/CLIP.git@ded190a052fdf4585bd685cee5bc96e0310d2c93
fastapi==0.135.1
filelock==3.25.2
fsspec==2026.2.0
ftfy==6.3.1
h11==0.16.0
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
packaging==26.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
regex==2026.2.28
setuptools==82.0.1
starlette==0.52.1
sympy==1.14.0
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
wcwidth==0.6.0
